#0: make web demo modules independant of pytest modules
