+++ models/demos/wormhole/yolov4/test_yolov4_performant_webdemo.py
@@ -4,12 +4,15 @@
 
 import pytest
 import ttnn
+import time
+import torch
 
 from models.utility_functions import run_for_wormhole_b0
-from models.demos.yolov4.tests.yolov4_perfomant import (
+from models.demos.yolov4.tests.yolov4_perfomant_webdemo import (
     run_yolov4_inference,
     run_yolov4_trace_inference,
     run_yolov4_trace_2cqs_inference,
+    Yolov4Trace2CQ,
 )
 
 
@@ -66,10 +69,27 @@ def test_run_yolov4_trace_2cqs_inference(
     enable_async_mode,
     model_location_generator,
 ):
-    run_yolov4_trace_2cqs_inference(
+    yolov4_trac2_2cq = Yolov4Trace2CQ()
+
+    yolov4_trac2_2cq.initialize_yolov4_trace_2cqs_inference(
         device,
         batch_size,
         act_dtype,
         weight_dtype,
-        model_location_generator,
+        model_location_generator=None,
     )
+    for iter in range(0, 10):
+        input_shape = (1, 3, 320, 320)
+        torch_input_tensor = torch.randn(input_shape, dtype=torch.float32)
+        n, c, h, w = torch_input_tensor.shape
+        torch_input_tensor = torch_input_tensor.permute(0, 2, 3, 1)
+        torch_input_tensor = torch_input_tensor.reshape(1, 1, h * w * n, c)
+        tt_inputs_host = ttnn.from_torch(torch_input_tensor, dtype=ttnn.bfloat16, layout=ttnn.ROW_MAJOR_LAYOUT)
+        tt_inputs_host = ttnn.pad(tt_inputs_host, [1, 1, n * h * w, 16], [0, 0, 0, 0], 0)
+
+        t0 = time.time()
+        output = yolov4_trac2_2cq.execute_yolov4_trace_2cqs_inference(tt_inputs_host)
+        t1 = time.time()
+        print("TIME", t1 - t0)
+
+    yolov4_trac2_2cq.release_yolov4_trace_2cqs_inference()
