+++ models/demos/yolov4/web_demo/server/requirements.txt
@@ -1,8 +1,5 @@
 fastapi==0.85.1
 uvicorn==0.19.0
 python-multipart==0.0.5
-transformers==4.20.1
-yolov5==7.0.11
-pytest==6.2.4
 
 -f https://download.pytorch.org/whl/cpu/torch_stable.html
